saving state.  more transaction functions.  Refactored HexStr/ParseHex functions into a crypto/utils file.
