+++ src/crypto/ark_crypto.cpp
@@ -2,6 +2,7 @@
 
 #include "constants/constants.h"
 #include "crypto/rfc6979.h"
+#include "crypto/util.h"
 #include "utilities/platform.h"
 #include "constants/networks.h"
 #include "models/transaction.h"
@@ -151,10 +152,6 @@ Account create_account(uint8_t network, const char* const passphrase) {
 	return Account(HexStr(pub_key).c_str(), address.c_str());
 }
 
-void get_transaction_bytes(const ARK::Transaction& transaction, uint8_t buffer[512]) {
-	
-}
-
 ARK::Transaction create_transaction(
 	uint8_t network, 
 	const char* const address, 
@@ -237,57 +234,4 @@ ARK::Transaction create_transaction(
 }
 
 }
-}
-
-namespace {
-
-const signed char p_util_hexdigit[256] =
-{ -1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
-0,1,2,3,4,5,6,7,8,9,-1,-1,-1,-1,-1,-1,
--1,0xa,0xb,0xc,0xd,0xe,0xf,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,0xa,0xb,0xc,0xd,0xe,0xf,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
--1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1, };
-
-}
-
-signed char HexDigit(char c)
-{
-	return p_util_hexdigit[(unsigned char)c];
-}
-
-std::vector<unsigned char> ParseHex(const char* psz)
-{
-	// convert hex dump to vector
-	std::vector<unsigned char> vch;
-	while (true)
-	{
-		while (isspace(*psz))
-			psz++;
-		signed char c = HexDigit(*psz++);
-		if (c == (signed char)-1)
-			break;
-		unsigned char n = (c << 4);
-		c = HexDigit(*psz++);
-		if (c == (signed char)-1)
-			break;
-		n |= c;
-		vch.push_back(n);
-	}
-	return vch;
-}
-
-std::vector<unsigned char> ParseHex(const std::string& str)
-{
-	return ParseHex(str.c_str());
 }
