+++ src/crypto/util.cpp
@@ -1,0 +1,54 @@
+#include "crypto/util.h"
+
+namespace {
+
+	const signed char p_util_hexdigit[256] =
+	{ -1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		0,1,2,3,4,5,6,7,8,9,-1,-1,-1,-1,-1,-1,
+		-1,0xa,0xb,0xc,0xd,0xe,0xf,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,0xa,0xb,0xc,0xd,0xe,0xf,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,
+		-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1, };
+
+}
+
+signed char HexDigit(char c)
+{
+	return p_util_hexdigit[(unsigned char)c];
+}
+
+std::vector<unsigned char> ParseHex(const char* psz)
+{
+	// convert hex dump to vector
+	std::vector<unsigned char> vch;
+	while (true)
+	{
+		while (isspace(*psz))
+			psz++;
+		signed char c = HexDigit(*psz++);
+		if (c == (signed char)-1)
+			break;
+		unsigned char n = (c << 4);
+		c = HexDigit(*psz++);
+		if (c == (signed char)-1)
+			break;
+		n |= c;
+		vch.push_back(n);
+	}
+	return vch;
+}
+
+std::vector<unsigned char> ParseHex(const std::string& str)
+{
+	return ParseHex(str.c_str());
+}
