+++ test/crypto/bip66.cpp
@@ -1,6 +1,7 @@
 #include "gtest/gtest.h"
 
 #include "crypto/ark_crypto.h"
+#include "crypto/util.h"
 
 TEST(crypto_bip66, toDER) {
 	auto r = ParseHex("33A69CD2065432A30F3D1CE4EB0D59B8AB58C74F27C41A7FDB5696AD4E6108C9"); // dec: 23362334225185207751494092901091441011938859014081160902781146257181456271561
